remove 'dummyviews' package, content of package moved directly to 'views' package
