+++ app/src/main/java/com/doplgangr/secrecy/views/CircularImageView.java
@@ -1,4 +1,4 @@
-package com.doplgangr.secrecy.views.dummyviews;
+package com.doplgangr.secrecy.views;
 
 import android.content.Context;
 import android.graphics.Bitmap;
