+++ app/src/main/res/layout/activity_main.xml
@@ -14,7 +14,7 @@
         app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
         app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
 
-    <com.doplgangr.secrecy.views.dummyviews.navdrawer.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    <com.doplgangr.secrecy.views.navdrawer.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:id="@+id/drawer_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
@@ -53,12 +53,12 @@
                     android:textColor="@color/white" />
             </RelativeLayout>
 
-            <com.doplgangr.secrecy.views.dummyviews.navdrawer.NavListView
+            <com.doplgangr.secrecy.views.navdrawer.NavListView
                 android:id="@+id/left_drawer_list"
                 android:layout_width="fill_parent"
                 android:layout_height="fill_parent"
                 android:background="@color/white" />
         </LinearLayout>
-    </com.doplgangr.secrecy.views.dummyviews.navdrawer.DrawerLayout>
+    </com.doplgangr.secrecy.views.navdrawer.DrawerLayout>
 
 </LinearLayout>
